Multigrid Projector 0.8.2
- Voxel collisions are now considered in block highlighting, they show up in cyan color. (client only)
- Block references (toolbar slots, turret rotors/hinges, etc.) are now properly restored on building blocks from projection. (client and server)
- Improved client only build in multiplayer if there is no MGP plugin on server side, including restoring block references. (client only build)
- Custom button names on button panels are now properly restored. Ported this fix from the Sections plugin.
- The "Fix Toolbars" projector action now restores all block references, not only toolbar slots.
- Minor stability fixes and code cleanup.

Since this plugin is pretty much complete, I'm planning on releasing MGP 1.0.0 in a few months.

The MGP API Agent code (for PB, Mod and other plugins) verifies the major version number of
the plugin and refuses to work it is not zero. With this MGP release I've removed this check,
so it will be possible to change the major version number to 1.

The agent code was copied into other scripts and mods to provide access to projection data.
If you copied the agent into your script or mod, then please update the code from the GitHub
repository (https://github.com/viktor-ferenczi/multigrid-projector/) according to your usage:

- **Script**: Copy the code from the `MGP API Agent` region of `IngameApiTest/Script/Script.cs`
- **Mod**: Copy the source files from the `MultigridProjectorApi/Api` directory
- **Plugin**: Same source files as with mods, but you get access to the provider with reflection

There has been no change to the actual API, so all your existing code will remain compatible.

Big thanks to @spacegt for contributing to the development of the above features.