Delete Plugins/Mods/HudLCDv2.xml
Despite its name saying "v2", this mod is actually an outdated fork of the other HudLCD mod by Jawa. This mod (hudlcd v2) causes crashes when using the nexus seamless plugin, and has fallen behind its predecessor in features/updates. 

I'm suggesting it be taken off the plugin hub due to its misleading name, the fact that it's no longer being maintained, and also to support the original author of the mod, Jawa.