Toolbar Manager 1.5.6
- Standardized configuration code
- Using the publicizer (no more reflections)
- Harmony 2.3.6 (to be in sync with Pulsar)

This release does not bring any player visible changes, other than a configuration change (see below). This release prepares the code for some new features, which may come at a later date.

If you are using a non-English keyboard layout, which requires the use of a hotkey other than pipe (`|`), then please re-select your hotkey in Toolbar Manager's configuration. Now you can combine the key with Ctrl, Shift or Alt as needed, which may be better in your case. There have been an incompatible change to the configuration code which made this necessary. Sorry for the inconvenience.